#25 refactor: 共通部品移動
